move SmallByteSlice into new utils submodule
